fix(billing): move the copy guards where CI actually runs them, and widen them
The number-pinning half of the new guard ran in NO CI job. `src/__tests__/`
is in neither Postgres step (ci.yml:222 `src/server src/lib`, :234 `src/app`)
and the unit job has no DATABASE_URL, so cap<->matrix, the organisation
allowance, the Pro Plus differentiator and the "largest grant" superlative
reported 6 pending / exit 0 — the exact failure vitest.config.ts's own header
documents. Moved to src/lib/__tests__/; measured in CI's own `src/lib`
command, the six now execute rather than skip.

Guards extracted to src/lib/copy-truth.ts, a plain module: exporting them
from a *.test.ts made every importer re-run the suite (a 4-test probe
reported 21; now 4). Tasks 3/4/7 import from there.

Both pattern lists were still beatable, and six measured rewordings are now
committed as tests: "an allowance of AI schedule runs for each division",
"AI scheduling: three runs a division", "a monthly quota of AI schedule
generations per division", "stays yours for as long as you want", "does not
lapse", "yours to keep". Two holes in the clause scoping closed as well —
dropping the colon made split(":")[0] return the whole description, and any
bare "active" satisfied a rule meant to assert a bound ("active immediately"
passed), so the bound's grammar is now required. Added the positive backstop
pro_plus lacked: a frame that survives with no recognised claim behind it is
a stale vocabulary, not a pass.

The extra-organisation rider was sold as "half the base rate" and is not:
$9 vs $19 is 47.4%, $19 vs $39 is 48.7%. It is NOT "a little under half"
either — in eur and aud on seazn_pro_monthly the halves fall on whole units
(1800/900, 2800/1400), so the rider is EXACTLY half and that wording would
have swapped one falsehood for another. Ships as "costs no more than half
the base rate", true in all twenty plan x interval x currency combinations,
with riderRateFaults tying the arithmetic to the claim's shape so either
side moving reds.
